+++ docs/api/conf.py
@@ -14,7 +14,7 @@ copyright = '2026, Interference Calculator Team'
 author = 'Interference Calculator Team'
 
 # The full version, including alpha/beta/rc tags
-release = '2.8.2'
+release = '2.8.3'
 version = '2.8'
 
 # -- General configuration ---------------------------------------------------